pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "receipt-designer"
version = "0.9.5"
description = "A desktop tool for designing and printing ESC/POS receipts."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE.md" }
authors = [{ name = "Josh Viles", email = "jviles5674@gmail.com" }]
keywords = ["escpos", "receipt", "designer", "PySide6", "thermal printer"]
dependencies = [
  "PySide6>=6.5",
  "Pillow>=9.0",
  "python-barcode>=0.15",
  "qrcode[pil]>=7.0",
  "pyserial>=3.5",
  "pyusb>=1.2",
  "python-escpos>=3.0",
]

[project.optional-dependencies]
treepoem = ["treepoem>=3.0"]
dev = ["pytest>=7.0"]

[project.urls]
Homepage = "https://github.com/ByteAndConquer/receipt-designer"
Issues = "https://github.com/ByteAndConquer/receipt-designer/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["receipt_designer*"]

[project.scripts]
receipt-designer = "receipt_designer.app:main"
